requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==4.4.2
docutils==0.19
ffmpeg==1.4
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.4.1
iso8601==1.1.0
isodate==0.6.1
jaraco.classes==3.2.3
keyring==23.13.1
ldap3==2.9.1
lxml==4.9.2
m3u8==3.4.0
markdown-it-py==2.2.0
mdurl==0.1.2
mega-pip==1.0
mega.py==1.0.8
more-itertools==9.1.0
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.2
pathlib==1.0.1
Pillow==9.5.0
pkginfo==1.9.6
proglog==0.1.10
pyasn1==0.4.8
pycountry==22.3.5
pycryptodome==3.17
pydirectory==0.1.15
Pygments==2.15.0
PyGPT==1.0.5
PySocks==1.7.1
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
pytube==12.1.2
pytubeX==0.1.4
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
streamlink==5.3.1
tenacity==5.1.5
tqdm==4.65.0
twine==4.0.2
twitch-dl==2.1.1
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
websocket-client==1.5.1
windows-curses==2.3.1
yarl==1.8.2
youtube-dl==2021.12.17
zipp==3.15.0
